+++ .travis.yml
@@ -18,6 +18,7 @@ before_script:
 services:
   - redis-server
 addons:
+  chrome: stable
   code_climate:
     repo_token: 15ed3ea3aa5e152f478928dbb01e5747253e54be2b90553170931dbc842428c7
 after_success:
